+++ bc/utils/wagtail_hooks.py
@@ -1,33 +1,34 @@
-from wagtail.contrib.modeladmin.options import (
-    ModelAdmin,
-    ModelAdminGroup,
-    modeladmin_register,
-)
+# Temporarily commented out while the news, events, and people apps are disabled
 
-from bc.events.models import EventType
-from bc.news.models import NewsType
-from bc.people.models import PersonType
+# from wagtail.contrib.modeladmin.options import (
+#     ModelAdmin,
+#     ModelAdminGroup,
+#     modeladmin_register,
+# )
 
-
-class EventTypeModelAdmin(ModelAdmin):
-    model = EventType
-    menu_icon = "tag"
+# from bc.events.models import EventType
+# from bc.news.models import NewsType
+# from bc.people.models import PersonType
 
 
-class NewsTypeModelAdmin(ModelAdmin):
-    model = NewsType
-    menu_icon = "tag"
+# class EventTypeModelAdmin(ModelAdmin):
+#     model = EventType
+#     menu_icon = "tag"
 
 
-class PersonTypeModelAdmin(ModelAdmin):
-    model = PersonType
-    menu_icon = "tag"
+# class NewsTypeModelAdmin(ModelAdmin):
+#     model = NewsType
+#     menu_icon = "tag"
 
 
-class TaxonomiesModelAdminGroup(ModelAdminGroup):
-    menu_label = "Taxonomies"
-    items = (NewsTypeModelAdmin, EventTypeModelAdmin, PersonTypeModelAdmin)
-    menu_icon = "tag"
+# class PersonTypeModelAdmin(ModelAdmin):
+#     model = PersonType
+#     menu_icon = "tag"
 
 
-modeladmin_register(TaxonomiesModelAdminGroup)
+# class TaxonomiesModelAdminGroup(ModelAdminGroup):
+#     menu_label = "Taxonomies"
+#     items = (NewsTypeModelAdmin, EventTypeModelAdmin, PersonTypeModelAdmin)
+#     menu_icon = "tag"
+
+# modeladmin_register(TaxonomiesModelAdminGroup)