+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- feature x
+- feature 21: disable events, news, people, and forms apps as we are not using them yet.
 
 ## 1.0 (yyyy-mm-dd)